# Reorders input points in Scattered Interpolation query