--- a/src/axom/mint/mesh/UnstructuredMesh.hpp
+++ b/src/axom/mint/mesh/UnstructuredMesh.hpp
@@ -707,7 +707,6 @@ public:
    */
   virtual IndexType getNumberOfCellNodes(IndexType cellID = 0) const final override
   {
-    using CardinalityPolicy = typename CellToNodeRelation::CardinalityPolicy;
     return m_cell_node_rel.CardinalityPolicy::size(cellID);
   }
 
@@ -887,7 +886,6 @@ public:
    */
   virtual IndexType getNumberOfFaceNodes(IndexType faceID = 0) const final override
   {
-    using CardinalityPolicy = typename FaceToNodeRelation::CardinalityPolicy;
     return m_face_node_rel.CardinalityPolicy::size(faceID);
   }