--- a/src/axom/mint/mesh/UnstructuredMesh.hpp
+++ b/src/axom/mint/mesh/UnstructuredMesh.hpp
@@ -707,7 +707,8 @@ public:
    */
   virtual IndexType getNumberOfCellNodes(IndexType cellID = 0) const final override
   {
-    return m_cell_node_rel.CardinalityPolicy::size(cellID);
+    using CardinalityPolicy = typename CellToNodeRelation::CardinalityPolicy;
+    return static_cast<CardinalityPolicy>(m_cell_node_rel).size(cellID);
   }
 
   /*!
@@ -886,7 +887,8 @@ public:
    */
   virtual IndexType getNumberOfFaceNodes(IndexType faceID = 0) const final override
   {
-    return m_face_node_rel.CardinalityPolicy::size(faceID);
+    using CardinalityPolicy = typename FaceToNodeRelation::CardinalityPolicy;
+    return static_cast<CardinalityPolicy>(m_face_node_rel).size(faceID);
   }
 
   /*!