--- a/src/axom/slam/IndirectionSet.hpp
+++ b/src/axom/slam/IndirectionSet.hpp
@@ -22,7 +22,7 @@ namespace axom
 namespace slam
 {
 /**
- * \brief Alias template for an OrderedSet with indirection over an array
+ * \brief Alias template for an OrderedSet with indirection over a C array
  *
  * \tparam PosType The position type for indexing into the set
  * \tparam ElemType The type for the set's elements
@@ -30,13 +30,13 @@ namespace slam
  */
 template <typename PosType = slam::DefaultPositionType,
           typename ElemType = slam::DefaultElementType>
-using ArrayIndirectionSet =
+using CArrayIndirectionSet =
   OrderedSet<PosType,
              ElemType,
              policies::RuntimeSize<PosType>,
              policies::ZeroOffset<PosType>,
              policies::StrideOne<PosType>,
-             policies::ArrayIndirection<PosType, ElemType>>;
+             policies::CArrayIndirection<PosType, ElemType>>;
 
 /**
  * \brief Alias template for an OrderedSet with indirection over an stl vector