--- a/src/axom/slam/IndirectionSet.hpp
+++ b/src/axom/slam/IndirectionSet.hpp
@@ -56,6 +56,23 @@ using VectorIndirectionSet =
              policies::STLVectorIndirection<PosType, ElemType>>;
 
 /**
+ * \brief Alias template for an OrderedSet with indirection over an axom::Array
+ *
+ * \tparam PosType The position type for indexing into the set
+ * \tparam ElemType The type for the set's elements
+ * \sa OrderedSet
+ */
+template <typename PosType = slam::DefaultPositionType,
+          typename ElemType = slam::DefaultElementType>
+using ArrayIndirectionSet =
+  OrderedSet<PosType,
+             ElemType,
+             policies::RuntimeSize<PosType>,
+             policies::ZeroOffset<PosType>,
+             policies::StrideOne<PosType>,
+             policies::ArrayIndirection<PosType, ElemType>>;
+
+/**
  * \brief Alias template for an OrderedSet with indirection over an axom::ArrayView
  *
  * \tparam PosType The position type for indexing into the set
@@ -71,23 +88,6 @@ using ArrayViewIndirectionSet =
              policies::ZeroOffset<PosType>,
              policies::StrideOne<PosType>,
              policies::ArrayViewIndirection<PosType, ElemType>>;
-
-/**
- * \brief Alias template for an OrderedSet with indirection over an axom::Array
- *
- * \tparam PosType The position type for indexing into the set
- * \tparam ElemType The type for the set's elements
- * \sa OrderedSet
- */
-template <typename PosType = slam::DefaultPositionType,
-          typename ElemType = slam::DefaultElementType>
-using CoreArrayIndirectionSet =
-  OrderedSet<PosType,
-             ElemType,
-             policies::RuntimeSize<PosType>,
-             policies::ZeroOffset<PosType>,
-             policies::StrideOne<PosType>,
-             policies::CoreArrayIndirection<PosType, ElemType>>;
 
 }  // end namespace slam
 }  // end namespace axom