# Summary

- This PR is a feature
- It reorders the input points for the scattered interpolation example following the Biased Randomized Insertion Order (BRIO) algorithm. This improves the theoretical expected runtime as well as the practical performance of the Delaunay mesh generation algorithm.
- BRIO was introduced in the following paper: 
   N. Amenta, S. Choi, and G. Rote. "Incremental constructions con BRIO." 
   Proceedings of the 19th annual symposium on Computational geometry, 2003.

- This PR also adds a function to `quest::ScatteredInterpolation` to directly get the interpolation indices and weights. This allows users to perform the interpolation themselves.
- The scattered_interpolation example uses the latter to compare the two interpolation schemes and to check interpolation for the position of the input points
- Finally, we add unit tests for the scattered_interpolation example
- Also renames some internal policy classes in slam now that we're using `axom::Array`. Specifically `slam::policies::ArrayIndirection` is now `slam::policies::CArrayIndirection` and `slam::policies::CoreArrayIndirection` is now `slam::policies::ArrayIndirection`.

### Performance improvement

To measure the improvement in this branch, I ran experiments in 2D and 3D on a series of randomly generated input and query points in the unit square/cube. 

I configured the code with the clang@10 compiler on a toss3 machine in LLNL's LC cluster using the `RelWithDebInfo` configuration, the `-march=native` compiler flag and the `-DAXOM_DEBUG_DEFINE=OFF` axom option (which removes asserts). I ran each experiment three times and took the average.

| With BRIO   | 2D         |            |            |            |    3D      |           |            |   |
|   ---            |  --:       |   --:      |  --:       |  --:       |  --:       |  --:      |  --:       |    --:         |
|  # input and query points | 10,000     | 100,000    | 1,000,000  | 10,000,000 | 10,000     | 100,000   | 1,000,000  | 10,000,000 |
| Insert input points | 270,460    | 277,592    | 277,323    | 264,894    | 51,833     | 52,358    | 52,643     | 51,655* |
| Locate query points | 1,542,527  | 657,566    | 313,124    | 96,237     | 950,174    | 512,505   | 247,778    | 113,028* |
| Interpolate fields  | 25,890,575 | 21,424,170 | 9,513,809  | 5,182,302  | 12,862,174 | 9,678,306 | 4,642,876  | 2,717,755* |
---
| Original (w/o BRIO) | 2D         |            |            |            |    3D      |           |            |                 |
|   ---            |  --:       |   --:      |  --:       |  --:       |  --:       |  --:      |  --:       |    --:         |
|  # input and query points | 10,000     | 100,000    | 1,000,000  | 10,000,000 | 10,000     | 100,000   | 1,000,000  | 10,000,000 |
| Insert input points | 253,648    | 218,569    | 150,548    | 70,897     | 48,285     | 46,622    | 36,502     | 28,346* |
| Locate query points | 1,642,490  | 777,691    | 295,070    | 82,584     | 1,094,842  | 631,869   | 255,445    | 125,962* |
| Interpolate fields  | 24,669,485 | 14,711,874 | 9,287,055  | 4,222,028  | 13,326,577 | 9,727,948 | 4,115,913  | 2,643,922* |

(*) Only one of the three runs were available for the 10,000,000 sample runs in 3D

As can be seen in the tables, reordering the input vertices using BRIO yields a uniform generation rate across different dataset sizes -- roughly an insertion rate of 270k pts/second in 2D and 50k pts/second in 3D.  Without BRIO, the rate was steadily declining as the mesh size increased.

Since the query points are not coherent, the query rates still reduce as the mesh gets larger.
I'm not entirely sure why the interpolation rates are reducing. I'd speculate that this is due to memory/cache coherence.

#### TODO
- [x] Update RELEASE-NOTES
- [x]  Post performance data