@kennyweiss 2022-06-30
@dtaller wrote:
> Would it be possible to add a function that would return a bounding box to the user? 

Sure -- I'll add it