@rhornung67 2022-06-30
@kennyweiss does the reordering have the potential to expose more parallelism? how does it impact data locality (I'm thinking in terms of GPU execution)?