@kennyweiss 2022-07-01
@rhornung67 wrote:
> @kennyweiss does the reordering have the potential to expose more parallelism? how does it impact data locality (I'm thinking in terms of GPU execution)?

When the points are inserted in a spatially coherent order, the search for the simplex containing the insertion point and the "conflicts" (simplices that need to be removed) are smaller.
You might be interested in the following papers, which focus on OpenMP parallelism using a similar Delaunay generation algorithm w/ BRIO -- https://doi.org/10.1016/j.cad.2016.07.018 and https://doi.org/10.1002/nme.5987